+++ src/dotnet/diagnostics-client-plugin/Counters/Producer/CounterProducerConfiguration.cs
@@ -22,26 +22,35 @@ internal sealed class CounterProducerConfiguration
         SessionId = sessionId;
         RefreshInterval = refreshInterval;
 
-        _counterProviders = new CounterProviderCollection(listOfCounterProviders);
-
-        var eventPipeProviderCount = listOfMetrics is not null
-            ? _counterProviders.Count() + 1
-            : _counterProviders.Count();
-        var eventPipeProviders = new List<EventPipeProvider>(eventPipeProviderCount);
-
-        var counterProviders = EventPipeProviderFactory.CreateCounterProviders(_counterProviders.Providers(), 
-            refreshInterval);
-        eventPipeProviders.AddRange(counterProviders);
-
-        if (listOfMetrics is not null)
+        if (string.IsNullOrEmpty(listOfCounterProviders) && string.IsNullOrEmpty(listOfMetrics))
         {
-            var metricCollection = new MetricCollection(listOfMetrics);
-            var metricProvider = EventPipeProviderFactory.CreateMetricProvider(sessionId, metricCollection.Metrics,
-                refreshInterval, maxTimeSeries, maxHistograms);
-            eventPipeProviders.Add(metricProvider);
+            _counterProviders = new CounterProviderCollection();
+            EventPipeProviders = EventPipeProviderFactory.CreateCounterProviders(_counterProviders.Providers(), 
+                refreshInterval);
         }
+        else
+        {
+            _counterProviders = new CounterProviderCollection(listOfCounterProviders);
 
-        EventPipeProviders = eventPipeProviders.AsReadOnly();
+            var eventPipeProviderCount = listOfMetrics is not null
+                ? _counterProviders.Count() + 1
+                : _counterProviders.Count();
+            var eventPipeProviders = new List<EventPipeProvider>(eventPipeProviderCount);
+
+            var counterProviders = EventPipeProviderFactory.CreateCounterProviders(_counterProviders.Providers(), 
+                refreshInterval);
+            eventPipeProviders.AddRange(counterProviders);
+
+            if (listOfMetrics is not null)
+            {
+                var metricCollection = new MetricCollection(listOfMetrics);
+                var metricProvider = EventPipeProviderFactory.CreateMetricProvider(sessionId, metricCollection.Metrics,
+                    refreshInterval, maxTimeSeries, maxHistograms);
+                eventPipeProviders.Add(metricProvider);
+            }
+
+            EventPipeProviders = eventPipeProviders.AsReadOnly();
+        }
     }
 
     internal bool IsCounterEnabled(string provider, string counter) => _counterProviders.Contains(provider, counter);