+++ src/dotnet/diagnostics-client-plugin/Counters/ValueCounter.cs
@@ -5,7 +5,9 @@ namespace DiagnosticsClientPlugin.Counters.Common;
 internal readonly record struct ValueCounter(
     DateTime TimeStamp,
     string Name,
+    string DisplayName,
     string ProviderName,
     double Value,
-    CounterType Type
+    CounterType Type,
+    string? Tags
 );