+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 - Support for 2022.2 version of the platform
 - Additional information about process
+- Collect and monitor System.Diagnostics.Metrics with counters
 
 ### Fixed
 - Trace provider parsing algorithm